Prevent NPE in Dependency.get() in BrightnessController
When com.android.systemui process died for some reason, the process will
be restarted and will schedule to restart services, such as
SystemUIService and KeyguardService, if start activity BrightnessDialog
before these services, then NPE will happen, the reason is that this
activity will use Dependency which is initalized in these services.

Init Dependency in BrightnessController before using it.

Change-Id: I83f490bf3380a79991c9219ee0238e77a164bf7c
CRs-Fixed: 2083957
PackageManagerService: Avoid NPE in matchComponentForVerifier

mRequiredVerifierPackage, which is passed as an argument
to matchComponentForVerifier, can be null - which causes
Null pointer Exception during string comparison.
Protect the call to matchComponentForVerifier with a null
check on mRequiredVerifierPackage.

Test: make successful

Change-Id: Ie4253c44833a5b82e8e5b4159321e5db104d0eb7
Signed-off-by: Shibin George <shibing@codeaurora.org>